+++ plugins/utilities/discord_richpresence.py
@@ -15,6 +15,7 @@ from pathlib import Path
 from os import getcwd, remove
 from bauiv1lib.popup import PopupWindow
 
+
 import asyncio
 import http.client
 import ast
@@ -113,9 +114,6 @@ MAPNAME_ID = {
     "android": "1097728392280932453",
 }
 ANDROID = babase.app.classic.platform == "android"
-APP_VERSION = _babase.app.version if build_number < 21282 else (
-    _babase.app.env.engine_version if build_number > 21823 else _babase.app.env.version)
-
 
 if ANDROID:  # !can add ios in future